First attempt at windows build

+++ include/galsim/Random.h
@@ -91,7 +91,7 @@ namespace galsim {
          *
          * @param[in] lseed A long-integer seed for the RNG.
          */
-        explicit BaseDeviate(long lseed);
+        explicit BaseDeviate(long long lseed);
 
         /**
          * @brief Construct a new BaseDeviate, sharing the random number generator with rhs.
@@ -144,7 +144,7 @@ namespace galsim {
          *
          * Note that this will reseed all Deviates currently sharing the RNG with this one.
          */
-        virtual void seed(long lseed);
+        virtual void seed(long long lseed);
 
         /**
          * @brief Like seed(lseed), but severs the relationship between other Deviates.
@@ -152,7 +152,7 @@ namespace galsim {
          * Other Deviates that had been using the same RNG will be unaffected, while this
          * Deviate will obtain a fresh RNG seed according to lseed.
          */
-        void reset(long lseed);
+        void reset(long long lseed);
 
         /**
          * @brief Make this object share its random number generator with another Deviate.
@@ -183,7 +183,7 @@ namespace galsim {
         /**
          * @brief Get a random value in its raw form as a long integer.
          */
-        long raw();
+        long long raw();
 
         /**
          * @brief Draw a new random number from the distribution
@@ -254,7 +254,7 @@ namespace galsim {
          *
          * @param[in] lseed A long-integer seed for the RNG.
          */
-        UniformDeviate(long lseed);
+        UniformDeviate(long long lseed);
 
         /// @brief Construct a new UniformDeviate, sharing the random number generator with rhs.
         UniformDeviate(const BaseDeviate& rhs);
@@ -312,7 +312,7 @@ namespace galsim {
          * @param[in] mean  Mean of the output distribution
          * @param[in] sigma Standard deviation of the distribution
          */
-        GaussianDeviate(long lseed, double mean, double sigma);
+        GaussianDeviate(long long lseed, double mean, double sigma);
 
         /**
          * @brief Construct a new Gaussian-distributed RNG, sharing the random number
@@ -422,7 +422,7 @@ namespace galsim {
          * @param[in] N Number of "coin flips" per trial
          * @param[in] p Probability of success per coin flip.
          */
-        BinomialDeviate(long lseed, int N, double p);
+        BinomialDeviate(long long lseed, int N, double p);
 
         /**
          * @brief Construct a new binomial-distributed RNG, sharing the random number
@@ -518,7 +518,7 @@ namespace galsim {
          * @param[in] lseed Seed to use
          * @param[in] mean  Mean of the output distribution
          */
-        PoissonDeviate(long lseed, double mean);
+        PoissonDeviate(long long lseed, double mean);
 
         /**
          * @brief Construct a new Poisson-distributed RNG, sharing the random number
@@ -616,7 +616,7 @@ namespace galsim {
          * @param[in] a    Shape parameter of the output distribution, must be > 0.
          * @param[in] b    Scale parameter of the distribution, must be > 0.
          */
-        WeibullDeviate(long lseed, double a, double b);
+        WeibullDeviate(long long lseed, double a, double b);
 
         /**
          * @brief Construct a new Weibull-distributed RNG, sharing the random number
@@ -719,7 +719,7 @@ namespace galsim {
          * @param[in] k      Shape parameter of the output distribution, must be > 0.
          * @param[in] theta  Scale parameter of the distribution, must be > 0.
          */
-        GammaDeviate(long lseed, double k, double theta);
+        GammaDeviate(long long lseed, double k, double theta);
 
         /**
          * @brief Construct a new Gamma-distributed RNG, sharing the random number
@@ -823,7 +823,7 @@ namespace galsim {
          * @param[in] lseed Seed to use
          * @param[in] n     Number of degrees of freedom for the output distribution, must be > 0.
          */
-        Chi2Deviate(long lseed, double n);
+        Chi2Deviate(long long lseed, double n);
 
         /**
          * @brief Construct a new Chi^2-distributed RNG, sharing the random number
