+++ pysrc/Random.cpp
@@ -49,11 +49,11 @@ namespace galsim {
     void pyExportRandom(py::module& _galsim)
     {
         py::class_<BaseDeviate> (_galsim, "BaseDeviateImpl")
-            .def(py::init<long>())
+            .def(py::init<long long>())
             .def(py::init<const BaseDeviate&>())
             .def(py::init<const char*>())
             .def("duplicate", &BaseDeviate::duplicate)
-            .def("seed", (void (BaseDeviate::*) (long) )&BaseDeviate::seed)
+            .def("seed", (void (BaseDeviate::*) (long long) )&BaseDeviate::seed)
             .def("reset", (void (BaseDeviate::*) (const BaseDeviate&) )&BaseDeviate::reset)
             .def("clearCache", &BaseDeviate::clearCache)
             .def("serialize", &BaseDeviate::serialize)
